vote_score as float

+++ models/user.js
@@ -90,8 +90,8 @@ const User = sequelize.define('User', {
     allowNull: true,
     default: 0
   },
-  vote_score: {
-    type: Sequelize.INTEGER,
+  vote_score: { // https://gitlab.com/watog-app/sql-nodejs/issues/11
+    type: Sequelize.FLOAT,
     allowNull: true,
     default: 0
   },
